Crate sparkfun_pro_micro

Board Support Crate for SparkFun Pro Micro.

This crate provides abstractions for interfacing with the hardware of SparkFun Pro Micro. It re-exports functionality from the underlying HAL in ways that make more sense for this particular board. For example, the pins are named by what is printed on the PCB instead of the MCU names.

Examples

A number of examples can be found in the examples/ subdirectory of this crate.

Getting Started

Please follow the guide from avr-hal's README for steps on how to set up a project with this board. A rough skeleton for an application looks like this:

#![no_std]
#![no_main]

// Pull in the panic handler from panic-halt
extern crate panic_halt;

// The prelude just exports all HAL traits anonymously which makes
// all trait methods available.  This is probably something that
// should always be added.
use sparkfun_pro_micro::prelude::*;

// Define the entry-point for the application.  This can only be
// done once in the entire dependency tree.
#[sparkfun_pro_micro::entry]
fn main() -> ! {
    // Get the peripheral singletons for interacting with them.
    let dp = sparkfun_pro_micro::Peripherals::take().unwrap();

    unimplemented!()
}

Re-exports

pub use atmega32u4_hal as hal;

Modules

adc
Support for the Analog to Digital Converter

pac
ATmega32U4

prelude

pwm
Support for PWM pins

spi
Support for the Serial Peripheral Interface

Structs

DDR
Generic DDR that works for all ports

Peripherals
All the peripherals

Pins
Reexport of the Pro Micro's pins, with the names they have on the board

Functions

delay_ms
Wait (busy spin) for ms milliseconds

delay_us
Wait (busy spin) for us microseconds

Type Definitions

Delay
Busy-Delay

I2cMaster
I2C Master on pins D2 (SDA) and D3 (SCL)

Serial
Serial (UART) interface on pins D0 (RX) and D1 (TX)

Attribute Macros

entry
See avr_device::entry.